IntimateTales_Roleplay/tests/test_data_synchronizer.py
```python
import unittest
from src.integration.data_synchronizer import DataSynchronizer

class TestDataSynchronizer(unittest.TestCase):

    def setUp(self):
        self.data_synchronizer = DataSynchronizer()

    def test_synchronizeData(self):
        # Test data
        story_data = {"title": "Test Story", "content": "This is a test story."}
        character_data = {"name": "Test Character", "traits": ["Brave", "Smart"]}

        # Synchronize data
        self.data_synchronizer.synchronizeData(story_data, character_data)

        # Check if data is synchronized
        self.assertEqual(self.data_synchronizer.story_data, story_data)
        self.assertEqual(self.data_synchronizer.character_data, character_data)

    def test_integrateWithCorePlugin(self):
        # Test data
        core_plugin_data = {"plugin_name": "Test Plugin", "version": "1.0.0"}

        # Integrate with core plugin
        self.data_synchronizer.integrateWithCorePlugin(core_plugin_data)

        # Check if data is integrated
        self.assertEqual(self.data_synchronizer.core_plugin_data, core_plugin_data)

if __name__ == '__main__':
    unittest.main()
```